Eddy Current Correction

This function calls eddy from FSL for DTI Processing

eddy(
  infile,
  mask,
  acq_file,
  index_file,
  bvecs,
  bvals,
  topup = NULL,
  outfile = NULL,
  retimg = TRUE,
  opts = "",
  verbose = TRUE,
  eddy_cmd = c("eddy", "eddy_openmp", "eddy_cuda"),
  ...
)

Arguments

infile
input filename of 4D image.

mask
Mask filename (or class nifti)

acq_file
A text-file describing the acquisition parameters for the different images in infile The format of this file is identical to that used by topup (though the parameter is called --datain there).

index_file
A text-file that determines the relationship between on the one hand the images in infile and on the other hand the acquisition parameters in acq_file.

bvecs
A text file with normalised vectors describing the direction of the diffusion weighting.

bvals
A text file with b-values describing the "amount of" diffusion weighting

topup
This should only be specified if you have previously run `topup` on your data and should be the same name that you gave as an argument to the --out parameter when you ran topup, aka the base name for output files from topup.

outfile
Output file basename

retimg
(logical) return image of class nifti

opts
Additional options to pass to arguments passed to eddy

verbose
print diagnostic messages

eddy_cmd
The version of eddy to run.

...
Not currently used

Value

Result from system command currently