Register using FNIRT, but doing Affine Registration as well

This function calls fnirt to register infile to reffile and either saves the image or returns an object of class nifti, but does the affine registration first

fnirt_with_affine(infile, reffile, flirt.omat = NULL, flirt.outfile = NULL,
  outfile = NULL, retimg = TRUE, reorient = FALSE, intern = FALSE,
  flirt.opts = "", opts = "", verbose = TRUE, ...)

Arguments

infile
(character) input filename

reffile
(character) reference image to be registered to

flirt.omat
(character) Filename of output affine matrix

flirt.outfile
(character) Filename of output affine-registered image

outfile
(character) output filename

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) pass to system

flirt.opts
(character) additional options to FLIRT

opts
(character) additional options to FNIRT

verbose
(logical) print out command before running

...
additional arguments passed to readnii.

Value

character or logical depending on intern