Form clusters, report information about clusters and/or perform cluster-based inference. Wrapper for cluster

Form clusters, report information about clusters and/or perform cluster-based inference. Wrapper for cluster

fsl_cluster(file, threshold, retimg = FALSE, reorient = FALSE,
  opts = "", cope_image = NULL, pthresh = NULL, peakdist = 0,
  volume = FALSE, smooth_est = NULL, voxel_resel = NULL,
  fractional = FALSE, connectivity = 26, mm = FALSE,
  find_minima = FALSE, standard_image = NULL, verbose = TRUE, ...)

fslcluster(..., retimg = TRUE)

read_cluster_table(file)

Arguments

file
filename of input volume

threshold
threshold for input volume

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

opts
(character) operations to be passed to cluster

cope_image
filename of input cope volume

pthresh
p-threshold

peakdist
minimum distance between local maxima/minima, in mm (default 0)

volume
number of voxels in the mask

smooth_est
smoothness estimate = sqrt(det(Lambda))

voxel_resel
Size of one resel in voxel units

fractional
interprets the threshold as a fraction of the robust range

connectivity
the connectivity of voxels (default 26)

mm
use mm, not voxel, coordinates

find_minima
find minima instead of maxima

standard_image
filename for standard-space volume

verbose
(logical) print out command before running

...
additional arguments to pass to fslcmd

Value

A list of filenames of outputs and tables:

opvalsfilename for image output of log pvals

oindexfilename for output of cluster index (in size order)

othreshfilename for output of thresholded image

olmaxfilename for output of local maxima text file

olmaximfilename for output of local maxima volume

osizefilename for output of size image

omaxfilename for output of max image

omeanfilename for output of mean image

Examples

if (have_fsl()) {
file = mni_fname(brain = TRUE, mask = FALSE)
threshold = 6000
clus = fsl_cluster(file, threshold)
}

#> FSLDIR='/usr/local/fsl'; PATH=${FSLDIR}/bin:${PATH};export PATH FSLDIR; sh "${FSLDIR}/etc/fslconf/fsl.sh"; FSLOUTPUTTYPE=NIFTI_GZ; export FSLOUTPUTTYPE; ${FSLDIR}/bin/cluster --in="/usr/local/fsl/data/standard/MNI152_T1_1mm_brain.nii.gz"  --thresh=6000 --oindex=/tmp/Rtmpt5qRjE/file7c2e51272e8a.nii.gz --othresh=/tmp/Rtmpt5qRjE/file7c2e592a4be1.nii.gz --olmax=/tmp/Rtmpt5qRjE/file7c2e3526aad3 --olmaxim=/tmp/Rtmpt5qRjE/file7c2e32fd2b2.nii.gz --osize=/tmp/Rtmpt5qRjE/file7c2e107f4c6e.nii.gz --omax=/tmp/Rtmpt5qRjE/file7c2e32f028ee.nii.gz --omean=/tmp/Rtmpt5qRjE/file7c2e5dfa931a.nii.gz  --minclustersize --verbose > /tmp/Rtmpt5qRjE/file7c2e3dd749b4 ;