Tool to deface a structural T1w image.

Tool to deface a structural T1w image.

fsl_deface(
  file,
  outfile = NULL,
  retimg = TRUE,
  opts = "",
  deface_cropped = FALSE,
  bet_fractional_intensity = NULL,
  bias_correct = FALSE,
  shift_xyz = NULL,
  cog_xyz = NULL,
  reorient = FALSE,
  intern = FALSE,
  verbose = TRUE,
  ...
)

Arguments

file
(character) input image to estimate edge strength

outfile
(character) resultant image name (optional)

retimg
(logical) return image of class nifti

opts
(character) operations to be passed to fsl_deface

deface_cropped
apply the defacing to the cropped image instead of the original image

bet_fractional_intensity
fractional intensity for bet (0->1); default=0.5;

bias_correct
Bias-correct the input image (with fast);

shift_xyz
Shift, in mm, x-, y- and z-directions, to shift face mask by;

cog_xyz
centre-of-gravity for bet (voxels, not mm);

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) to be passed to system

verbose
print diagnostic messages

...
additional arguments passed to fslcmd.

Examples

# \donttest{
if (have_fsl()) {
  file = mni_fname(mm = 1, brain = FALSE)
  out = fsl_deface(file, retimg = FALSE)
}
# }