FSL Change file type

This function calls fslchfiletype

fslchfiletype(file, filetype = "NIFTI_GZ", outfile = NULL, retimg = TRUE,
  reorient = FALSE, intern = FALSE, verbose = TRUE, ...)

Arguments

file
(character) image to be manipulated

filetype
filetype to change image to

outfile
Output filename. If NULL, will overwrite input file

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) to be passed to system

verbose
(logical) print out command before running

...
additional arguments passed to readnii.

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.