Register EPI images to Structural image

This function calls epi_reg, designed to register EPI images (typically functional or diffusion) to structural (e.g. T1-weighted) image.

fslepi_reg(
  epi,
  t1,
  t1_brain,
  outfile = NULL,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  fmap = NULL,
  fmap_mag = NULL,
  fmap_mag_brain = NULL,
  echo_spacing = NA,
  phase_enc_dir = c("x", "y", "z", "-x", "-y", "-z"),
  weight = NULL,
  verbose = TRUE,
  opts = "",
  ...
)

fsl_epi_reg(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

epi_reg(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

Arguments

epi

EPI image, character or nifti object

t1

whole head T1 image , character or nifti object

t1_brain

brain extracted T1 image

outfile

output registered image filename

retimg

(logical) return image of class nifti

reorient

(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern

(logical) to be passed to system

fmap

fieldmap image (in rad/s)

fmap_mag

fieldmap magnitude image - whole head extracted

fmap_mag_brain

fieldmap magnitude image - brain extracted

echo_spacing

Effective EPI echo spacing (sometimes called dwell time) - in seconds

phase_enc_dir

phase encoding direction, dir = x/y/z/-x/-y/-z

weight

weighting image (in T1 space)

verbose

(logical) print out command before running

opts

(character) operations to be passed to fslmaths

...

additional arguments passed to readnii.

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.

Note

Functions with underscores have different defaults and will return an output filename, so to be used for piping