Modulus Remainder of 2 Images using FSL

This function calls fslmaths -rem. The R functions wraps fslmaths

fsl_rem(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

fslrem(
  file,
  file2,
  outfile = NULL,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  opts = "",
  ...
)

Arguments

...
additional arguments passed to readnii.

outfile
(character) resultant image name (optional)

retimg
(logical) return image of class nifti

file
(character) input image

file2
(character) image to divide the current image by and take remainder

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) to be passed to system

opts
(character) operations to be passed to fslmaths

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.

Note

Functions with underscores have different defaults and will return an output filename, so to be used for piping