FSL ROI

This function calls fslroi

fslroi(file, xmin = 0, xsize = -1, ymin = 0, ysize = -1, zmin = 0,
  zsize = -1, tmin = NULL, tsize = NULL, outfile = NULL,
  retimg = TRUE, reorient = FALSE, intern = FALSE, verbose = TRUE, ...)

fslroi_time(file, tmin = NULL, tsize = NULL, ...)

Arguments

file
(character) image to be manipulated

xmin
Minimum index for x-dimension

xsize
Size of ROI in x-dimension

ymin
Minimum index for y-dimension

ysize
Size of ROI in y-dimension

zmin
Minimum index for z-dimension

zsize
Size of ROI in z-dimension

tmin
Minimum index for t-dimension

tsize
Size of ROI in t-dimension

outfile
(character) resultant image name (optional)

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) to be passed to system

verbose
(logical) print out command before running

...
additional arguments passed to readnii.

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.

Note

Indexing (in both time and space) starts with 0 not 1! Inputting -1 for a size will set it to the full image extent for that dimension.