FSL Stats Help

This function calls fslstats's help

fslstats.help()

Value

Prints help output and returns output as character vector

Examples

if (have.fsl()){
 fslstats.help()
}

#> Usage: fslstats [preoptions] <input> [options]preoption -t will give a separate output line for each 3D volume of a 4D timeseriespreoption -K < indexMask > will generate seperate n submasks from indexMask, for indexvalues 1..n where n is the maximum index value in indexMask, and generate statistics for each submaskNote - options are applied in order, e.g. -M -l 10 -M will report the non-zero mean, apply a threshold and then report the new nonzero mean-l <lthresh> : set lower threshold-u <uthresh> : set upper threshold-r           : output <robust min intensity> <robust max intensity>-R           : output <min intensity> <max intensity>-e           : output mean entropy ; mean(-i*ln(i))-E           : output mean entropy (of nonzero voxels)-v           : output <voxels> <volume>-V           : output <voxels> <volume> (for nonzero voxels)-m           : output mean-M           : output mean (for nonzero voxels)-s           : output standard deviation-S           : output standard deviation (for nonzero voxels)-w           : output smallest ROI <xmin> <xsize> <ymin> <ysize> <zmin> <zsize> <tmin> <tsize> containing nonzero voxels-x           : output co-ordinates of maximum voxel-X           : output co-ordinates of minimum voxel-c           : output centre-of-gravity (cog) in mm coordinates-C           : output centre-of-gravity (cog) in voxel coordinates-p <n>       : output nth percentile (n between 0 and 100)-P <n>       : output nth percentile (for nonzero voxels)-a           : use absolute values of all image intensities-n           : treat NaN or Inf as zero for subsequent stats-k <mask>    : use the specified image (filename) for masking - overrides lower and upper thresholds-d <image>   : take the difference between the base image and the image specified here-h <nbins>   : output a histogram (for the thresholded/masked voxels only) with nbins-H <nbins> <min> <max>   : output a histogram (for the thresholded/masked voxels only) with nbins and histogram limits of min and maxNote - thresholds are not inclusive ie lthresh<allowed<uthresh