Subsample image by factor of 2

This function calls fslmaths -subsamp2 to subsample an image and either saves the image or returns an object of class nifti

fsl_sub2(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

fslsub2(file, outfile = NULL, retimg = TRUE, reorient = FALSE,
  intern = FALSE, verbose = TRUE, ...)

Arguments

...
additional arguments passed to readnii.

outfile
(character) name of resultant subsampled file

retimg
(logical) return image of class nifti

file
(character) filename of image to be subsampled

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) pass to system

verbose
(logical) print out command before running

Value

character or logical depending on intern

Note

Functions with underscores have different defaults and will return an output filename, so to be used for piping

Examples

if (have.fsl()){
system.time({
x = array(rnorm(1e6), dim = c(100, 100, 100))
img = nifti(x, dim= c(100, 100, 100),
datatype = convert.datatype()$FLOAT32, cal.min = min(x),
cal.max = max(x), pixdim = rep(1, 4))
subsamp = fslsub2(img, retimg=TRUE)
print(voxdim(subsamp))
})
}

#> fslmaths "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/Rtmpw5j0CC/file5c234f304b3.nii.gz"  -subsamp2 "/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//Rtmpw5j0CC/file5c23eb99eed";

#> [1] 2 2 2

#>    user  system elapsed 
#>   1.393   0.283   1.711