Threshold an image

This function calls fslmaths -thr -uthr to threshold an image and either saves the image or returns an object of class nifti

fsl_thresh(..., outfile = tempfile(fileext = ".nii.gz"), retimg = FALSE)

fslthresh(
  file,
  outfile = NULL,
  thresh = 0,
  uthresh = NULL,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  opts = "",
  verbose = TRUE,
  ...
)

Arguments

...

additional arguments passed to readnii.

outfile

(character) name of resultant thresholded file

retimg

(logical) return image of class nifti

file

(character) filename of image to be thresholded

thresh

(numeric) threshold (anything below set to 0)

uthresh

(numeric) upper threshold (anything above set to 0)

reorient

(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern

(logical) pass to system

opts

(character) additional options to be passed to fslmaths

verbose

(logical) print out command before running

Value

character or logical depending on intern

Note

Functions with underscores have different defaults and will return an output filename, so to be used for piping

Examples

if (have.fsl()){
system.time({
x = array(rnorm(1e6), dim = c(100, 100, 100))
img = nifti(x, dim= c(100, 100, 100), 
datatype = convert.datatype()$FLOAT32, cal.min = min(x), 
cal.max = max(x), pixdim = rep(1, 4))
thresh = fslthresh(img, thresh=0, uthresh = 2, retimg=TRUE)
})
} 
#> fslmaths "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpKzob0N/file85756581b897.nii.gz" -thr 0.000000 -uthr 2.000000  "/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpKzob0N/file85751879d0e6"
#>    user  system elapsed 
#>   0.646   0.041   0.689