Get Q and S Forms of orientation matrix

This function obtains the s and q forms of an image transformation matrix

getForms(file, verbose = FALSE, ...)

Arguments

file

(character) filename of image to pass to header

verbose

(logical) passed to fslhd

...

options passed to checkimg

Value

list with elements of sform and qform and their respective codes

Examples

if (have.fsl()){
 mnifile = mni_fname("2")
 getForms(mnifile)
}   
#> $qform
#>      [,1] [,2] [,3] [,4]
#> [1,]   -2    0    0   90
#> [2,]    0    2    0 -126
#> [3,]    0    0    2  -72
#> [4,]    0    0    0    1
#> 
#> $sform
#>      [,1] [,2] [,3] [,4]
#> [1,]   -2    0    0   90
#> [2,]    0    2    0 -126
#> [3,]    0    0    2  -72
#> [4,]    0    0    0    1
#> 
#> $sor
#> [1] "sform_xorient\tRight-to-Left"        
#> [2] "sform_yorient\tPosterior-to-Anterior"
#> [3] "sform_zorient\tInferior-to-Superior" 
#> 
#> $qor
#> [1] "qform_xorient\tRight-to-Left"        
#> [2] "qform_yorient\tPosterior-to-Anterior"
#> [3] "qform_zorient\tInferior-to-Superior" 
#> 
#> $ssor
#> [1] "RL" "PA" "IS"
#> 
#> $sqor
#> [1] "RL" "PA" "IS"
#> 
#> $sform_code
#> [1] 4
#> 
#> $qform_code
#> [1] 4
#>