Probabilistic diffusion tractography with multiple fibre orientations

This function wraps probtrackx from FSL

probtrackx(
  samples = "merged",
  mask,
  seed,
  outdir = "fdt_paths",
  verbose = TRUE,
  mode = NULL,
  targetmasks = NULL,
  mask2 = NULL,
  waypoints = NULL,
  network = FALSE,
  mesh = NULL,
  seedref = NULL,
  dir = FALSE,
  forcedir = FALSE,
  opd = FALSE,
  pd = FALSE,
  os2t = FALSE,
  avoid = NULL,
  stop = NULL,
  xfm = NULL,
  invxfm = NULL,
  nsamples = 5000,
  nsteps = 2000,
  distthresh = 0,
  cthr = 0.2,
  fibthresh = 0.01,
  sampvox = FALSE,
  steplength = 0.5,
  loopcheck = FALSE,
  usef = FALSE,
  randfib = c(0, 1, 2, 3),
  fibst = 1,
  modeuler = FALSE,
  rseed = NULL,
  s2tastext = FALSE,
  opts = ""
)

Arguments

samples

(nifti/character) Basename for samples files

mask

(nifti/character) Bet binary mask file in diffusion space

seed

(nifti/character) Seed volume, or voxel, or ascii file with multiple volumes, or freesurfer label file

outdir

(character) Output file (default='fdt_paths')

verbose

(logical/numeric) Verbose level, [0-2]

mode

(character) Use --mode=simple for single seed voxel

targetmasks

(character) File containing a list of target masks - required for seeds_to_targets classification

mask2

(nifti/character) Second mask in twomask_symm mode.

waypoints

(nifti/character) Waypoint mask or ascii list of waypoint masks - only keep paths going through ALL the masks

network

(logical) Activate network mode - only keep paths going through at least one seed mask (required if multiple seed masks)

mesh

(character) Freesurfer-type surface descriptor (in ascii format)

seedref

(nifti/character) Reference vol to define seed space in simple mode - diffusion space assumed if absent

dir

(logical) Directory to put the final volumes in - code makes this directory - default='logdir'

forcedir

(logical) Use the actual directory name given - i.e. don't add + to make a new directory

opd

(logical) Output path distribution

pd

(logical) Correct path distribution for the length of the pathways

os2t

(logical) Output seeds to targets

avoid

(nifti/character) Reject pathways passing through locations given by this mask

stop

(nifti/character) Stop tracking at locations given by this mask file

xfm

(character) Transform taking seed space to DTI space (either FLIRT matrix or FNIRT warpfield) - default is identity

invxfm

(character) Transform taking DTI space to seed space (compulsory when using a warpfield for seeds_to_dti)

nsamples

(numeric) Number of samples - default=5000

nsteps

(numeric) Number of steps per sample - default=2000

distthresh

(numeric) Discards samples shorter than this threshold (in mm - default=0)

cthr

(numeric) Curvature threshold - default=0.2

fibthresh

(numeric) Volume fraction before subsidiary fibre orientations are considered - default=0.01

sampvox

(logical) Sample random points within seed voxels

steplength

(numeric) Steplength in mm - default=0.5

loopcheck

(logical) Perform loopchecks on paths - slower, but allows lower curvature threshold

usef

(logical) Use anisotropy to constrain tracking

randfib

(numeric) Default 0. Set to 1 to randomly sample initial fibres (with f > fibthresh). Set to 2 to sample in proportion fibres (with f>fibthresh) to f. Set to 3 to sample ALL populations at random (even if f<fibthresh)

fibst

(numeric) Force a starting fibre for tracking - default=1, i.e. first fibre orientation. Only works if randfib==0

modeuler

(logical) Use modified euler streamlining

rseed

(numeric) Random seed

s2tastext

(logical) Output seed-to-target counts as a text file (useful when seeding from a mesh)

opts

Additional options or way to specify things instead of command line arguments

Value

A filename of the output file