Run FIRST All

Wrapper for run_first_all from FSL for FIRST analysis segmentation of subcortical structures

run_first_all(img, oprefix, brain_extracted = FALSE, structures = NULL,
  affine = NULL, opts = "", verbose = TRUE)

Arguments

img

pecifies the input image (T1-weighted)

oprefix

specifies the output image basename (extensions will be added to this)

brain_extracted

specifies that the input image has been brain extracted

structures

a restricted set of structures to be segmented

affine

specifies the affine registration matrix to standard space (optional)

opts

(character) operations to be passed to run_first_all

verbose

(logical) print out command before running

Value

Result of system