Bayesian Estimation of Diffusion Parameters Obtained using Sampling Techniques with Crossing Fibers

Calls xfibres from FSL to fit, also known as bedpostx

xfibres(infile, bvecs, bvals, mask = NULL, nfibres = 1, bet.opts = "",
  verbose = TRUE, njumps = NULL, burnin = NULL, burnin_noard = NULL,
  sampleevery = NULL, updateproposalevery = NULL, seed = NULL,
  noard = FALSE, allard = FALSE, nospat = FALSE, nonlinear = FALSE,
  cnonlinear = FALSE, rician = FALSE, f0 = FALSE, ardf0 = FALSE,
  opts = "")

Arguments

infile
Input filename

bvecs
b-vectors: matrix of 3 columns or filename of ASCII text file

bvals
b-values: vector of same length as number of rows of b-vectors or filename of ASCII text file

mask
Mask filename

nfibres
Maximum number of fibres to fit in each voxel (default 1)

bet.opts
Options for fslbet if mask is not supplied

verbose
print diagnostic messages

njumps
num of jumps to be made by MCMC (default is 5000)

burnin
Total num of jumps at start of MCMC to be discarded (default is 0)

burnin_noard
num of burnin jumps before the ard is imposed (default is 0)

sampleevery
num of jumps for each sample (MCMC) (default is 1)

updateproposalevery
num of jumps for each update to the proposal density std (MCMC) (default is 40)

seed
for pseudo random number generator

noard
Turn ARD off on all fibres

allard
Turn ARD on on all fibres

nospat
Initialise with tensor, not spatially

nonlinear
Initialise with nonlinear fitting

cnonlinear
Initialise with constrained nonlinear fitting

rician
Use Rician noise modelling

f0
Add to the model an unattenuated signal compartment

ardf0
Use ard on f0

opts
Additional options for xfibres. There should not be any left out in the current arguments, but opts may be a way some prefer to input options.

Value

Output from system