DTI Fitting Procedure from FSL

Calls dtifit from FSL

dtifit(infile, bvecs, bvals, mask = NULL, outprefix = NULL, opts = "",
  bet.opts = "", verbose = TRUE, sse = FALSE, save_tensor = FALSE)

Arguments

infile

Input filename

bvecs

b-vectors: matrix of 3 columns or filename of ASCII text file

bvals

b-values: vector of same length as number of rows of b-vectors or filename of ASCII text file

mask

Mask filename

outprefix

Output prefix

opts

Additional options for dtifit

bet.opts

Options for fslbet if mask is not supplied

verbose

print diagnostic messages

sse

Save sum of squared errors

save_tensor

Save tensor file out

Value

Vector of character filenames of output. See Note

Note

On successful completion of the command, the following files will be output, which are: mask - the mask used in the analysis outprefix_V1 - 1st eigenvector outprefix_V2 - 2nd eigenvector outprefix_V3 - 3rd eigenvector outprefix_L1 - 1st eigenvalue outprefix_L2 - 2nd eigenvalue outprefix_L3 - 3rd eigenvalue outprefix_MD - mean diffusivity outprefix_FA - fractional anisotropy outprefix_MO - mode of the anisotropy (oblate ~ -1; isotropic ~ 0; prolate ~ 1) outprefix_S0 - raw T2 signal with no diffusion weighting optional output If sse = TRUE, then the additional file will be present: outprefix_sse - Sum of squared error If save_tensor = TRUE, then the additional file will be present: outprefix_tensor - tensor as a 4D file in this order: Dxx,Dxy,Dxz,Dyy,Dyz,Dzz