Run FIRST All Help

This function calls run_first_all's help

run_first_all.help()

Value

Prints help output and returns output as character vector

Examples

library(fslr)

if (have.fsl()){
 run_first_all.help()
}

#> Usage: run_first_all [options] -i <input_image> -o <output_image>Optional arguments:  -m <method>      : method must be one of auto, fast, none or a (numerical) threshold value  -b               : input is already brain extracted  -s <name>        : run only on one specified structure (e.g. L_Hipp) or a comma separated list (no spaces)  -a <img2std.mat> : use affine matrix (do not re-run registration)  -3               : use 3-stage affine registration (only currently for hippocampus)  -d               : do not cleanup image output files (useful for debugging)  -v               : verbose output  -h               : display this help messagee.g.:  run_first_all -i im1 -o output_name