FSL SUSAN noise reduction

Implements Smallest Univalue Segment Assimilating Nucleus (SUSAN) noise reduction technique from FSL

susan(
  file,
  outfile = NULL,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  bthresh = 0.1,
  sigma = 3,
  dimg = c(3, 2),
  use_median = FALSE,
  n_usans = c(0, 1, 2),
  extra.scans = list(),
  opts = "",
  verbose = TRUE,
  ...
)

Arguments

file
(character) image to be manipulated

outfile
(character) resultant image name (optional)

retimg
(logical) return image of class nifti

reorient
(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern
(logical) to be passed to system

bthresh
brightness threshold and should be greater than noise level and less than contrast of edges to be preserved.

sigma
spatial size (sigma i.e. half-width) of smoothing in mm.

dimg
dimensionality (2 or 3) depending on whether smoothing is to be within-plane (2) or fully 3D (3).

use_median
determines whether to use a local median filter in the cases where single-point noise is detected (0 or 1).

n_usans
determines whether the smoothing area (USAN) is to be found from secondary images (0 1 or 2).

extra.scans
List of extra scans for USAN. List of n_usans elements, where each element has 2 named objects bthresh and filename

opts
(character) operations to be passed to susan, not currently used.

verbose
(logical) print out command before running

...
additional arguments passed to fslcmd.

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.

References

S.M. Smith and J.M. Brady. SUSAN -a new approach to low level image processing. International Journal of Computer Vision, 23(1):45-78, May 1997.