topup - calling FSL topup

A tool for estimating and correcting susceptibility induced distortions

topup(imain, datain, out = NULL, fout = NULL, iout = NULL,
  logout = NULL, warpres = 10, subsamp = 1, fwhm = 8, config = NULL,
  miter = 5, lambda = NULL, ssqlambda = 1, regmod = c("bending_energy",
  "membrane_energy"), estmov = 1, minmet = c(0, 1), splineorder = c(3, 2),
  numprec = c("double", "float"), interp = c("spline", "linear"),
  scale = c(0, 1), regrid = c(0, 1), verbose = TRUE)

fsl_topup(...)

Arguments

imain

name of 4D file with images

datain

name of text file with PE directions/times

out

base-name of output files (spline coefficients (Hz) and movement parameters)

fout

name of image file with field (Hz)

iout

name of 4D image file with unwarped images

logout

Name of log-file

warpres

(approximate) resolution (in mm) of warp basis for the different sub-sampling levels, default 10

subsamp

sub-sampling scheme, default 1

fwhm

FWHM (in mm) of gaussian smoothing kernel, default 8

config

Name of config file specifying command line arguments

miter

Max # of non-linear iterations, default 5

lambda

Weight of regularisation, default depending on ssqlambda and regmod switches. See user documetation.

ssqlambda

If set (=1), lambda is weighted by current ssq, default 1

regmod

Model for regularisation of warp-field [membrane_energy bending_energy], default bending_energy

estmov

Estimate movements if set, default 1 (true)

minmet

Minimisation method 0=Levenberg-Marquardt, 1=Scaled Conjugate Gradient, default 0 (LM)

splineorder

Order of spline, 2->Qadratic spline, 3->Cubic spline. Default=3

numprec

Precision for representing Hessian, double or float. Default double

interp

Image interpolation model, linear or spline. Default spline

scale

If set (=1), the images are individually scaled to a common mean, default 0 (false)

regrid

If set (=1), the calculations are done in a different grid, default 1 (true)

verbose

Print diagonostic information while running

...

arguments passed to topup if using fsl_topup